Move sagas.js into own module (for docs)
