+++ lib/trace/sagas/index.js
@@ -3,8 +3,8 @@ const debug = debugModule("debugger:trace:sagas");
 
 import { take, takeEvery, put, select } from "redux-saga/effects";
 
-import * as actions from "./actions";
-import trace from "./selectors";
+import * as actions from "../actions";
+import trace from "../selectors";
 
 export function *waitForTrace() {
   let {steps} = yield take(actions.SAVE_STEPS);
